# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Focus Room Virtual Study Space API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
  error?: string;
}

export interface User {
  id: string;
  username: string;
  email: string;
  avatar?: string | null;
  theme?: string;
  studyGoalMinutes?: number;
  createdAt: string;
  totalFocusMinutes: number;
  currentStreak: number;
  longestStreak: number;
}

export interface RegisterRequest {
  /**
   * @minLength 3
   * @maxLength 30
   */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UpdateProfileRequest {
  username?: string;
  avatar?: string | null;
  theme?: string;
  studyGoalMinutes?: number;
}

export type TimerStatePhase =
  (typeof TimerStatePhase)[keyof typeof TimerStatePhase];

export const TimerStatePhase = {
  focus: "focus",
  break: "break",
  idle: "idle",
} as const;

export interface TimerState {
  phase: TimerStatePhase;
  isRunning: boolean;
  /** Seconds remaining */
  timeRemaining: number;
  startedAt?: string | null;
  pomodoroCount: number;
}

export interface Room {
  id: string;
  name: string;
  description?: string | null;
  category: string;
  isPrivate: boolean;
  hasPassword: boolean;
  maxParticipants: number;
  participantCount: number;
  background: string;
  focusDuration: number;
  breakDuration: number;
  hostId: string;
  hostUsername: string;
  timerState?: TimerState;
  createdAt: string;
}

export interface Participant {
  userId: string;
  username: string;
  avatar?: string | null;
  joinedAt: string;
  isHost: boolean;
}

export type RoomDetail = Room & {
  participants: Participant[];
};

export interface CreateRoomRequest {
  /**
   * @minLength 3
   * @maxLength 60
   */
  name: string;
  description?: string | null;
  category?: string;
  isPrivate?: boolean;
  password?: string | null;
  /**
   * @minimum 2
   * @maximum 50
   */
  maxParticipants?: number;
  background?: string;
  /**
   * @minimum 5
   * @maximum 90
   */
  focusDuration?: number;
  /**
   * @minimum 1
   * @maximum 30
   */
  breakDuration?: number;
}

export interface JoinRoomRequest {
  password?: string | null;
}

export type TimerUpdateRequestAction =
  (typeof TimerUpdateRequestAction)[keyof typeof TimerUpdateRequestAction];

export const TimerUpdateRequestAction = {
  start: "start",
  pause: "pause",
  reset: "reset",
  skip: "skip",
} as const;

export type TimerUpdateRequestPhase =
  | (typeof TimerUpdateRequestPhase)[keyof typeof TimerUpdateRequestPhase]
  | null;

export const TimerUpdateRequestPhase = {
  focus: "focus",
  break: "break",
} as const;

export interface TimerUpdateRequest {
  action: TimerUpdateRequestAction;
  phase?: TimerUpdateRequestPhase;
}

export interface StudySession {
  id: string;
  userId: string;
  roomId?: string | null;
  roomName?: string | null;
  startTime: string;
  endTime?: string | null;
  durationMinutes?: number | null;
  pomodorosCompleted: number;
  completed: boolean;
}

export interface CreateSessionRequest {
  roomId?: string | null;
}

export interface CompleteSessionRequest {
  pomodorosCompleted?: number;
}

export interface DashboardStats {
  todayMinutes: number;
  weeklyMinutes: number;
  monthlyMinutes: number;
  totalMinutes: number;
  totalSessions: number;
  completedSessions: number;
  currentStreak: number;
  longestStreak: number;
  studyGoalMinutes: number;
  goalProgressPercent: number;
  topProductiveHour?: number | null;
  weeklyTarget: number;
}

export interface ActivityDay {
  date: string;
  minutes: number;
  sessions: number;
}

export interface LeaderboardEntry {
  rank: number;
  userId: string;
  username: string;
  avatar?: string | null;
  focusMinutes: number;
  sessionsCompleted: number;
  currentStreak: number;
  isCurrentUser: boolean;
}

export type ListRoomsParams = {
  search?: string;
  category?: string;
};

export type ListSessionsParams = {
  limit?: number;
  offset?: number;
};

export type GetActivityParams = {
  days?: number;
};

export type GetLeaderboardParams = {
  period?: GetLeaderboardPeriod;
  limit?: number;
};

export type GetLeaderboardPeriod =
  (typeof GetLeaderboardPeriod)[keyof typeof GetLeaderboardPeriod];

export const GetLeaderboardPeriod = {
  daily: "daily",
  weekly: "weekly",
  monthly: "monthly",
  alltime: "alltime",
} as const;