# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Focus Room Virtual Study Space API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get user dashboard stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /dashboard/activity:
    get:
      operationId: getActivity
      tags: [dashboard]
      summary: Get weekly activity data for graphs
      security:
        - bearerAuth: []
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 7
      responses:
        "200":
          description: Activity data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityDay"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
      required:
        - message
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        avatar:
          type: string
          nullable: true
        theme:
          type: string
          default: dark
        studyGoalMinutes:
          type: integer
          default: 120
        createdAt:
          type: string
          format: date-time
        totalFocusMinutes:
          type: integer
          default: 0
        currentStreak:
          type: integer
          default: 0
        longestStreak:
          type: integer
          default: 0
      required:
        - id
        - username
        - email
        - createdAt
        - totalFocusMinutes
        - currentStreak
        - longestStreak
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required:
        - username
        - email
        - password
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    UpdateProfileRequest:
      type: object
      properties:
        username:
          type: string
        avatar:
          type: string
          nullable: true
        theme:
          type: string
        studyGoalMinutes:
          type: integer
...
    DashboardStats:
      type: object
      properties:
        todayMinutes:
          type: integer
        weeklyMinutes:
          type: integer
        monthlyMinutes:
          type: integer
        totalMinutes:
          type: integer
        totalSessions:
          type: integer
        completedSessions:
          type: integer
        currentStreak:
          type: integer
        longestStreak:
          type: integer
        studyGoalMinutes:
          type: integer
        goalProgressPercent:
          type: number
        topProductiveHour:
          type: integer
          nullable: true
        weeklyTarget:
          type: integer
          default: 600
      required:
        - todayMinutes
        - weeklyMinutes
        - monthlyMinutes
        - totalMinutes
        - totalSessions
        - completedSessions
        - currentStreak
        - longestStreak
        - studyGoalMinutes
        - goalProgressPercent
        - weeklyTarget
    ActivityDay:
      type: object
      properties:
        date:
          type: string
          format: date
        minutes:
          type: integer
        sessions:
          type: integer
      required:
        - date
        - minutes
        - sessions
...