# lib/api-zod/src/generated/types/registerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Focus Room Virtual Study Space API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterRequest {
  /**
   * @minLength 3
   * @maxLength 30
   */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}