# lib/api-zod/src/generated/types/updateProfileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Focus Room Virtual Study Space API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProfileRequest {
  username?: string;
  avatar?: string | null;
  theme?: string;
  studyGoalMinutes?: number;
}